* develop: (24 commits)
  UI fixups (#892)
  Update Public Suffix List data (#888)
  Use remembered credential even if it is empty (#880)
  Reset SSH passphrase after SSH key import (#885)
  Prevent cached passwords from being wiped (#884)
  build: uprev all dependencies (#882)
  github: update actions for updated branching logic
  Use a custom sshj config (#878)
  Remove API 30 from pull request test matrix (#879)
  Add Google Play/F-Droid badges to the README (#877)
  Retire Android Arsenal badge (#876)
  Add paragraphs to F-Droid summary (#874)
  Replace YAML with HTML in F-Droid fastlane summary (#873)
  Add relnotes for #871 (#872)
  Add org.gnu.icecat as a trusted multi-origin browser (#871)
  README: re-add F-Droid (#870)
  build: prepare next development version
  build: bump version to 1.9.0
  Prepare release 1.9.0
  PasswordCreationActivity: properly guard rename code
  ...

+++ app/src/main/java/com/zeapo/pwdstore/crypto/PasswordCreationActivity.kt
@@ -18,6 +18,7 @@ import androidx.lifecycle.lifecycleScope
 import com.github.ajalt.timberkt.e
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.zeapo.pwdstore.PasswordEntry
+import com.zeapo.pwdstore.utils.isInsideRepository
 import com.zeapo.pwdstore.R
 import com.zeapo.pwdstore.autofill.oreo.AutofillPreferences
 import com.zeapo.pwdstore.autofill.oreo.DirectoryStructure
@@ -235,6 +236,11 @@ class PasswordCreationActivity : BasePgpActivity(), OpenPgpServiceConnection.OnB
                                 snackbar(message = getString(R.string.password_creation_duplicate_error))
                                 return@executeApiAsync
                             }
+
+                            if (!isInsideRepository(file)) {
+                                snackbar(message = getString(R.string.message_error_destination_outside_repo))
+                                return@executeApiAsync
+                            }
                             try {
                                 file.outputStream().use {
                                     it.write(outputStream.toByteArray())
@@ -280,7 +286,7 @@ class PasswordCreationActivity : BasePgpActivity(), OpenPgpServiceConnection.OnB
                                 }
                             }
 
-                            if (category.isVisible && category.isEnabled) {
+                            if (category.isVisible && category.isEnabled && oldFileName != null) {
                                 val oldFile = File("$repoPath/${oldCategory?.trim('/')}/$oldFileName.gpg")
                                 if (oldFile.path != file.path && !oldFile.delete()) {
                                     setResult(RESULT_CANCELED)
